feat: adicionando middleware de validação na criação de assistidos
